+++ changelog.md
@@ -48,5 +48,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Save button in ModelizerNavigationBar
 - Add onUpdateFile function in ModelizerTextView and create FileTabHeader component
 - Add Reference type attributes management in ComponentDetailPanel
+- Add Unreferenced attributes management in ComponentDetailPanel
 
 [unreleased]: https://github.com/ditrit/leto-modelizer/blob/main/changelog.md#unreleased